refactor: Export core-database-postgres typedefs refactor: Remove un-used SPV.connection property refactor: Remove ConnectionInterface.connection property. This available in the 'options' object that's passed via constructor. refactor: Use inline-initialization instead of constructor initialization for concise code, unless we're accessing constructor parameters.
