+++ packages/core-database-postgres/src/spv.ts
@@ -3,8 +3,9 @@ const { Transaction } = models;
 
 import { app } from "@arkecosystem/core-container";
 import { AbstractLogger } from "@arkecosystem/core-logger";
-import {PostgresConnection} from "./connection";
+import { PostgresConnection } from "./connection";
 import { queries } from "./queries";
+import { QueryExecutor } from "./sql/query-executor";
 
 const logger = app.resolvePlugin<AbstractLogger>("logger");
 const config = app.getConfig();
@@ -12,14 +13,12 @@ const config = app.getConfig();
 const genesisWallets = config.get("genesisBlock.transactions").map(tx => tx.senderId);
 
 export class SPV {
-    private connection: any;
     private models: any;
     private walletManager: any;
-    private query: any;
+    private query: QueryExecutor;
     private activeDelegates: [];
 
-    constructor(connectionInterface : PostgresConnection) {
-        this.connection = connectionInterface.connection;
+    constructor(connectionInterface: PostgresConnection) {
         this.models = connectionInterface.models;
         this.walletManager = connectionInterface.walletManager;
         this.query = connectionInterface.query;
